Upgrade commons-codec to 1.14
Affecting commons-codec:commons-codec artifact, versions [,1.13)

commons-codec:commons-codec is a package that contains simple encoder and decoders for various formats such as Base64 and Hexadecimal.

Affected versions of this package are vulnerable to Information Exposure. When there is no byte array value that can be encoded into a string the Base32 implementation does not reject it, and instead decodes it into an arbitrary value which can be re-encoded again using the same implementation. This allows for information exposure exploits such as tunneling additional information via seemingly valid base 32 strings.

Upgrade commons-codec:commons-codec to version 1.13 or higher.

* [GitHub Commit](https://github.com/apache/commons-codec/commit/48b615756d1d770091ea3322eefc08011ee8b113)
* [Jira Issue](https://issues.apache.org/jira/browse/CODEC-134)